lsp: Refactor diagnostic method registration
Due to timing issues between the background sphinx build and
`textDocument/diagnostic` requests from the client, the only way to
provide a good experience is to issue `workspace/diagnostic/refresh`
requests each time a build completes to force the client to ask for an
update.

Taking advantage of the better `initialize` request handling in
`pygls` v2, we only register our pull-model diagnostic support when
the client advertises all of the capabilities we need to provide a
good experience.